Update EIP association doc (#18)
The current version of this doesn't seem to be working. It looks like there's a bug in the way `$${}` works in heredocs. After fixing that, I noticed that the script could be simplified to just use the AWS CLI instead of a python library.

Feel free to close if I missed something here.